Fix some build script error, can set version when making deb file, force rustls-pemfile to 1.x

+++ src/build.py
@@ -30,8 +30,12 @@ if build_result != 0:
 target_dir = os.path.join(temp_dir, "rust_build", project_name,"x86_64-unknown-linux-musl")
 print(f'build success at: {target_dir}')
 
+sdk_build_dir = os.path.join(build_dir, "kernel/buckyos_sdk")
+sdk_build_cmd = f'cd {sdk_build_dir} && npm i && npm run build'
+os.system(sdk_build_cmd)
+
 vite_build_dir = os.path.join(build_dir, "kernel/node_active")
-vite_build_cmd = f'cd {vite_build_dir} && npm run build'
+vite_build_cmd = f'cd {vite_build_dir} && npm i && npm run build'
 os.system(vite_build_cmd)
 
 print(f'npm build success at: {vite_build_dir}')
@@ -73,8 +77,9 @@ shutil.copy(os.path.join(build_dir, "killall.py"), destination_dir)
 src_dir = os.path.join(vite_build_dir, "dist")
 destination_dir = os.path.join(build_dir, "rootfs/bin/active")
 print(f'copying vite build {src_dir} to {destination_dir}')
-shutil.rmtree(destination_dir)
-shutil.copytree(src_dir, destination_dir)
+if os.path.exists(destination_dir):
+    shutil.rmtree(destination_dir)
+shutil.copytree(src_dir, destination_dir, dirs_exist_ok=True)
 print('copying files to rootfs & web3_bridge done')
